+++ bindings/rust/build.rs
@@ -2,15 +2,18 @@ use std::env;
 use std::path::PathBuf;
 
 fn main() {
-    let lib_dir = env::var_os("RUCKIG_C_LIB_DIR").map(PathBuf::from).unwrap_or_else(|| {
-        let manifest_dir = PathBuf::from(env::var("CARGO_MANIFEST_DIR").expect("CARGO_MANIFEST_DIR"));
-        manifest_dir
-            .join("..")
-            .join("..")
-            .join("out")
-            .join("build")
-            .join("windows-clang-ninja")
-    });
+    let lib_dir = env::var_os("RUCKIG_C_LIB_DIR")
+        .map(PathBuf::from)
+        .unwrap_or_else(|| {
+            let manifest_dir =
+                PathBuf::from(env::var("CARGO_MANIFEST_DIR").expect("CARGO_MANIFEST_DIR"));
+            manifest_dir
+                .join("..")
+                .join("..")
+                .join("out")
+                .join("build")
+                .join("windows-clang-ninja")
+        });
 
     println!("cargo:rustc-link-search=native={}", lib_dir.display());
     println!("cargo:rustc-link-lib=static=ruckig_c");